// src/api/generated/model/ambassadorControllerGetEventApplicationsParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { AmbassadorControllerGetEventApplicationsStatus } from './ambassadorControllerGetEventApplicationsStatus';

export type AmbassadorControllerGetEventApplicationsParams = {
/**
 * Фильтр по статусу заявки
 */
status?: AmbassadorControllerGetEventApplicationsStatus;
/**
 * Массив ID событий для фильтрации
 */
eventIds?: string[];
/**
 * Номер страницы (от 1)
 */
page?: number;
/**
 * Размер страницы
 */
size?: number;
};